requirements.txt
requests~=2.31.0
deepdiff~=6.7.1
selectolax~=0.3.17
SQLAlchemy~=2.0.36
python-dotenv~=0.20.0